Set TTYPath for user-session@.service
In order for a login session (managed by systemd-logind) to become fully
active, it must be assigned to both a seat and a VT.  Without using a
display manager, two options are available for ensuring these
assignments are made: (a) set XDG_SEAT and XDG_VTNR in the PAM session,
or (b) set PAM_TTY in the PAM session.

This commit implements the second option by setting TTYPath to
"/dev/tty1" for user-session@.service, thereby forcing systemd to set
PAM_TTY in the PAM session, which then sets XDG_SEAT to "seat0" and
XDG_VTNR to "1" for the user session.  "/dev/tty1" is the TTY of choice
because xorg-launch-helper spawns Xorg on tty1.